+++ src/combinator/ops.rs
@@ -3,3 +3,4 @@
 pub mod add;
 pub mod bitor;
 pub mod mul;
+pub mod not;